Modified the config file so it can take different ways of initializing the FENE parameters as either free parameters or in terms of physical quantities. Need to test to make sure working.
Changed order of dimensional splitting of full solve. Now it is 1) Solve
for U and P, 2) Solve for f in configuration space, 3) Advect f in
physical space, 4) Calculate S from f.

Started rewriting cell centroid calculator, because it is currently
broken. For the time being, code is using a totally bogus ad-hoc method
to ensure that forces are computed for points inside the circular
domain. Need to finish writing this.