Added support in the Geometry and Circle classes for storing the number and coordinates of the vertices of each cell. This greatly simplified the implementation of a method to calculate the centroid of each cell, which was also fixed and reintroduced into the solver (where the centroid of each cell is used to evaluate the spring force in that cell).
Added method called advectFromFlat to the FENEMaterialDerivative
UniformAdvector class. This just advects each point in configuration
space as though it is constant. Probably not the fastest way to do
advection, but first test is to make sure it works. It also adaptively
picks a CFL number of around 0.9 whenever possible, which guarantees
stability of the method.