This commit is barely different than the last. All day was spent on a bug in which creating a CellDoubleArray, FaceDoubleArray, or VertexDoubleArray from one of the new object oriented (pointer-based) grids would occasionally cause MATLAB to hang if the array was initialized and resized from within another function besides the MEX function. Looking at the actual stack trace of MATLAB, it was determined that the program was hanging on the destructor for the MemoryBlock<double> class, which is part of Blitz. This bug does not occur with identical C++ code run from the command line.
My best guess is that something is not quite right in the way MATLAB
handles memory management in C++ mex calls. Because the Grid class
uses dynamic memory management via the new and delete keywords, there
is a lot of potential for shennanigans to happen, and apparently they
do.

For now, the workaround is as follows: in a mex function, instead of
calling g->makeCellDoubleArray() to create a new array, declare the
array explicitly via:
	CellDoubleArray c;
	c.resize(g->xRange,g->yRange);
	c = 0;
For whatever reason, declaring the array directly inside the mex
function seems not to cause the error. It will remain to be seen
whether arrays that are created and destroyed entirely within other
routines (e.g. during a multigrid solve) will be destroyed correctly
or not.

God damnit that was a long day of debugging only to find out it is an
interaction between Blitz and MATLAB and my code that I can't really
change.