Added a piecewise constant interpolator. The Johansen Colella paper suggests this is good enough for multigrid, but I am very dubious.
Starting to add a templated OperatorFactory class for automating
creation of operators on different grids. May also want to create a
factory object for time/space operators, so that the same factory can
create the left hand and right hand side operators for something like
Crank-Nicholson. Major hiccup for today, just learned, is that class
templates require the full function definitions in header files, not in
a separate source file.