Improved, hopefully completed, OperatorFactory. Also added another class template called SplitOperatorFactory for left hand side and right hand side operators. This might need to be generalized further to a doubly templated object, e.g. OperatorFactory<OperatorType,numberOfOperators> so that more complicated time splittings, like BDF2, can be put inside factories.
Added a Crank Nicholson Factory. Next up is testing this with
multigrid. The Crank Nicholson Factory requires a value of deltaT,
which is set at initialization and is hard-coded into all the
operators generated by the factory. Somewhere down the line it might
be better to have things like a multiplicative constant separated from
the rest of the operator, so that it could change. The hard work of
generating the stencil is mostly in creating the boundary stencil, so
that should only be done once even if other constants change. For now,
this goes on the long term todo list.